+++ Example/PaymentSheet Example/PaymentSheet Example/PaymentSheetTestPlaygroundSettings.swift
@@ -157,6 +157,12 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
         case on
         case off
     }
+    enum PaymentMethodSave: String, PickerEnum {
+        static var enumName: String { "PaymentMethodSave" }
+
+        case enabled
+        case disabled
+    }
     enum PaymentMethodRemove: String, PickerEnum {
         static var enumName: String { "PaymentMethodRemove" }
 
@@ -174,6 +180,7 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
 
         case always
         case limited
+        case unspecified
         case unspecified_limited_always
         case notSet
 
@@ -183,6 +190,8 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
                 return ["always"]
             case .limited:
                 return ["limited"]
+            case .unspecified:
+                return ["unspecified"]
             case .unspecified_limited_always:
                 return ["unspecified", "limited", "always"]
             case .notSet:
@@ -382,6 +391,7 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
     var applePayEnabled: ApplePayEnabled
     var applePayButtonType: ApplePayButtonType
     var allowsDelayedPMs: AllowsDelayedPMs
+    var paymentMethodSave: PaymentMethodSave
     var paymentMethodRemove: PaymentMethodRemove
     var paymentMethodRedisplay: PaymentMethodRedisplay
     var paymentMethodAllowRedisplayFilters: PaymentMethodAllowRedisplayFilters
@@ -419,6 +429,7 @@ struct PaymentSheetTestPlaygroundSettings: Codable, Equatable {
             applePayEnabled: .on,
             applePayButtonType: .buy,
             allowsDelayedPMs: .on,
+            paymentMethodSave: .enabled,
             paymentMethodRemove: .enabled,
             paymentMethodRedisplay: .enabled,
             paymentMethodAllowRedisplayFilters: .always,