+++ StripePayments/StripePayments/Source/API Bindings/Models/Shared/STPIntentActionRedirectToURL.swift
@@ -19,6 +19,31 @@ public class STPIntentActionRedirectToURL: NSObject {
     /// authenticating.
     @objc public let returnURL: URL?
 
+    /// If true, we'll follow one 30x redirect and open the webview using the resulting URL.
+    @_spi(STP) public var followRedirects: Bool {
+        return stripeFlagValue("followRedirectsInSDK")
+    }
+
+    /// If true, we'll use ASWebAuthenticationSession instead of SFSafariViewController.
+    /// Some payment methods benefit from sharing cookies with Safari and
+    /// using URL protocol handlers to return to the merchant's app.
+    @_spi(STP) public var useWebAuthSession: Bool {
+        return stripeFlagValue("useWebAuthSession")
+    }
+
+    /// Checks the value of an internal Stripe URL flag, returning `true` if the flag exists and is true, otherwise returning `false`.
+    internal func stripeFlagValue(_ flagName: String) -> Bool {
+        if let urlComponents = URLComponents(url: url, resolvingAgainstBaseURL: false),
+           let host = urlComponents.host,
+           host.hasSuffix("stripe.com"),
+            let queryItems = urlComponents.queryItems,
+            let item = queryItems.first(where: { $0.name == flagName }),
+           item.value == "true" {
+            return true
+        }
+        return false
+    }
+
     @objc public let allResponseFields: [AnyHashable: Any]
 
     let threeDSSourceID: String?